==> 01-subreddit-scraper-template.ipynb <==
# install necessary packages
# pip install praw
# pip install datetime
# pip install pandas

# import necessary packages
import praw
import pandas as pd
import datetime

# optional package for debugging (helps chatgpt help you debug)
# import logging
# logging.basicConfig(level=logging.DEBUG)

# initialize reddit via PRAW
reddit = praw.Reddit(
    client_id="", # must create reddit developer acc first
    client_secret="", # copy and paste client_id and client_secret here
    user_agent="call-lang-evol-1.0 by u/book_nook_3000" # name of user_agent by u/username
)

# Check if script has correct scope
try:
    print("Scopes available:", reddit.auth.scopes())
except Exception as e:
    print("An error occurred:", e)
# if the result of this cell is {'*'}, you're ready to scrape!

# select subreddit to scrape
subreddit = reddit.subreddit("")

# Scraping the top posts of all time
posts = subreddit.top(time_filter = "all", limit = None)
 
posts_dict = {"Title": [], "Post Text": [],
              "ID": [], "Score": [], "Upvote Ratio": [],
              "Total Comments": [],"Created On":[], "Post URL": [],
              "Original Content": []
              }

start_date = '01-01-20 00:00:00' # change start date if necessary
start_date = datetime.datetime.strptime(start_date, '%d-%m-%y %H:%M:%S').timestamp()

for post in posts:
    # Date of each posts' creation
    date = post.created_utc
    if date > start_date:
        # Title of each post
        posts_dict["Title"].append(post.title)
     
        # Text inside a post
        posts_dict["Post Text"].append(post.selftext)
     
        # Unique ID of each post
        posts_dict["ID"].append(post.id)
     
        # The score of a post
        posts_dict["Score"].append(post.score)
        
        # Upvote Ratio of a post
        posts_dict["Upvote Ratio"].append(post.upvote_ratio)
     
        # Total number of comments inside the post
        posts_dict["Total Comments"].append(post.num_comments)
         
        # Date the post was Created
        posts_dict["Created On"].append(post.created_utc)
        
        # URL of each post
        posts_dict["Post URL"].append(post.url)
        
        # Flair of each post
        posts_dict["Original Content"].append(post.is_original_content)
        
# Saving the data in a pandas dataframe
all_posts = pd.DataFrame(posts_dict)
all_posts['Created On'] = pd.to_datetime(all_posts['Created On'],  unit='s')
all_posts

# save raw data file to csv, place in /raw data folder
all_posts.to_csv('/Users/sophiachung/Documents/reddit_proj/raw/r_[subredddit]_dataset.csv')
