==> 02-subreddit-tidier-template.ipynb <==
# import packages
import pandas as pd

# overview of function
# inputs: name of dataset
# returns: new file path, original shape (num posts), final shape (num posts)

def subreddit_tidier(subreddit_name):
    # read in raw csv
    path = 'raw/r_' + subreddit_name + '_dataset.csv'
    df = pd.read_csv(path)
    print('original size of df:', df.shape[0])

    # remove any posts that are NaN for column 'Post Text'
    df = df[df['Post Text'].notna()]
    print('size of df after removing NaN:', df.shape[0])

    # discard any posts that include “EDIT: “ or “edit: “
    # messes with the focus on evolution / timing of the post
    df = df[~(df['Post Text'].str.contains('EDIT:') | 
              df['Post Text'].str.contains('edit:') | 
              df['Post Text'].str.contains('Edit:'))]
    print('size of df after removing edited posts:', df.shape[0])

    # set cutoff date
    df = df[df['Created On'] < '2024-09-01 00:00:00']
    print('size of df after tidying:', df.shape[0])

    # sort by column 'Created On'
    df = df.sort_values(by=['Created On'], ignore_index=True)
    df = df.rename(columns={'Unnamed: 0': 'Popularity Index'})

    # save new tidied dataset
    output_file_path = 'tidied/r_' + subreddit_name + '_tidied.csv'
    df.to_csv(output_file_path, index=False)
    print("\ncleaned dataset saved to:", output_file_path)

subreddit_tidier('printedcircuitboard')

subreddits = ['adops'] # create list of subreddits to scrape

for subreddit in subreddits: 
    name = subreddit.lower()
    subreddit_tidier(name)
