==> 03-subreddit-tpp-template.ipynb <==
# import packages
import pandas as pd
import numpy as np
import re
import string
import nltk
from nltk.corpus import stopwords
nltk.download('stopwords')
from nltk.stem import WordNetLemmatizer
nltk.download('wordnet')

import re

# function to remove all types of URLs
def remove_urls(text):
    # regex pattern to match URLs starting with [https://...] or (https://...)
    # handles [https://...] and (https://...)
    url_pattern = r'\[https?://[^\]]+\]|\(https?://[^\)]+\)'  
    
    # regex pattern to match URLs without http/https, just domains
    domain_pattern = r'\b\w+\.com\b|\b\w+\.co\b|\b\w+\.store\b|\b\w+\.org\b|\b\w+\.edu\b|\b\w+\.net\b|\b\w+\.us\b|\b\w+\.uk\b|\b\w+\.eu\b|\b\w+\.ca\b'
    
    # combine both patterns and remove all matches
    combined_pattern = f'{url_pattern}|{domain_pattern}'
    
    cleaned_text = re.sub(combined_pattern, '', text)  # replace the URL with an empty string
    return cleaned_text

string.punctuation[6]

# defining the function to remove apostrophes, specifically
def remove_punctuation(text):
        punctuationfree="".join([i for i in text if i not in string.punctuation[6]])
        return punctuationfree

def tokenization(text):
    tokens = re.split(r'\W+',text)
    return tokens

stopwords = nltk.corpus.stopwords.words('english')
# defining the function to remove stopwords from tokenized text
def remove_stopwords(text):
    output= [i for i in text if i not in stopwords]
    return output

# defining the object for Lemmatization
wordnet_lemmatizer = WordNetLemmatizer()
# defining the function for lemmatization
def lemmatizer(text):
    lemm_text = [wordnet_lemmatizer.lemmatize(word) for word in text]
    return lemm_text

def text_preprocessing(subreddit_name):
    path = 'tidied/r_' + subreddit_name + '_tidied.csv'
    df = pd.read_csv(path)
    df = df[['ID', 'Created On', 'Post Text']]
    df = df.rename(columns = {'Post Text': 'text'})
    
    # applying function to the column with the Reddit text
    df['no_urls'] = df['text'].apply(lambda x: remove_urls(x))
    
    # applying the function to remove punctuation 
    df['clean_msg']= df['no_urls'].apply(lambda x:remove_punctuation(x))
   
    # lowering text
    df['msg_lower']= df['clean_msg'].apply(lambda x: x.lower())
    
    # applying tokenization to text
    df['msg_tokenized']= df['msg_lower'].apply(lambda x: tokenization(x))

    # remove stopwords
    df['no_stopwords']= df['msg_tokenized'].apply(lambda x:remove_stopwords(x))
    
    # applying lemmatization to text
    df['msg_lemmatized']=df['no_stopwords'].apply(lambda x:lemmatizer(x))

    # save new pre-processed dataset
    output_file_path = 'tpp/r_' + subreddit_name + '_tpp.csv'
    df.to_csv(output_file_path, index=False)
    print("\ntext-preprocessed dataset saved to:", output_file_path)

# test full function
subreddit_name = 'adops'
text_preprocessing(subreddit_name)

subreddits = ['adops'] # create list of subreddits to pre-process

for subreddit in subreddits: 
    name = subreddit.lower()
    text_preprocessing(name)

# trying something
def text_preprocessing(subreddit_name):
    path = 'tidied/r_' + subreddit_name + '_tidied.csv'
    df = pd.read_csv(path)
    df = df[['ID', 'Created On', 'Post Text']]
    df = df.rename(columns = {'Post Text': 'text'})
    
    # Applying function to the column with the Reddit text
    df['no_urls'] = df['text'].apply(lambda x: remove_urls(x))
   
    # Lowering text
    df['msg_lower']= df['no_urls'].apply(lambda x: x.lower())
    
    # Applying tokenization to text
    df['msg_tokenized']= df['msg_lower'].apply(lambda x: tokenization(x))

    # Remove stopwords
    df['no_stopwords']= df['msg_tokenized'].apply(lambda x:remove_stopwords(x))
    
    # Applying lemmatization to text
    df['msg_lemmatized']=df['no_stopwords'].apply(lambda x:lemmatizer(x))

    # save new pre-processed dataset
    output_file_path = 'tpp/r_' + subreddit_name + '_tpp.csv'
    df.to_csv(output_file_path, index=False)
    print("\ntext-preprocessed dataset saved to:", output_file_path)

# test full function
subreddit_name = 'roborock'
text_preprocessing(subreddit_name)
