==> 05-subreddit-sentiment-analysis-template.ipynb <==
import numpy as np
import pandas as pd
from textblob import TextBlob
import nltk
nltk.download('brown')

def sentiment_analysis(subreddit):
    # read in the raw data file
    path = 'raw/r_' + subreddit + '_dataset.csv'
    df = pd.read_csv(path)

    # pull text from posts
    df = df[['Title', 'Post Text']].dropna()
    df = df.reset_index(drop=True)

    # combine text from title and post
    df['Text'] = df['Title'] + ' ' + df['Post Text']

    # initialize new columns
    df['Tags'] = ''
    df['Noun Phrases'] = ''

    # analyze average polarity of each text cell
    for post in range(len(df)):
        blob = TextBlob(df.at[post, 'Text'])
        df.at[post,'Tags'] = blob.tags
        df.at[post,'Noun Phrases'] = blob.noun_phrases
        polarity = []
        for sentence in blob.sentences:
            polarity.append(sentence.sentiment.polarity)
        if polarity: 
            df.at[post,'Polarity'] = np.mean(polarity)
        else:
            df.at[post,'Polarity'] = np.nan

    output_file_path = 'sentiment/r_' + subreddit + '_sentiment.csv'
    df.to_csv(output_file_path, index=True)
    print("\nsentiment analysis dataframe saved to:", output_file_path)

subreddit_list = ['adops']

for subreddit in subreddit_list:
    name = subreddit.lower()
    sentiment_analysis(name)
